fix(owo): refactor and check in DMs properly
